+++ src/main/java/com/ofss/repository/StockRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface StockRepository extends CrudRepository<Stock, Integer>{
+public interface StockRepository extends CrudRepository<Stock, Long>{
 }
